feedback/requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pyparsing==3.0.7
pysrt==1.1.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.8
#psycopg2-binary==2.9.3
